Add solutions for Q101 and Q102 problems
Added Q101.c and Q102.c, each implementing a C program for array search problems: Q101 finds the first and last occurrence of a target in a sorted array, and Q102 finds the index of the ceil of a target value. Corresponding compiled executables Q101.exe and Q102.exe are also included.